# Conflicts:
#	mala/targets/target_base.py

+++ mala/datahandling/data_converter.py
@@ -5,8 +5,8 @@ import numpy as np
 
 from mala.common.parallelizer import printout, get_rank
 from mala.common.parameters import ParametersData
-from mala.descriptors.descriptor_interface import DescriptorInterface
-from mala.targets.target_interface import TargetInterface
+from mala.descriptors.descriptor import Descriptor
+from mala.targets.target import Target
 
 
 class DataConverter:
@@ -20,22 +20,22 @@ class DataConverter:
     parameters : mala.common.parameters.Parameters
         The parameters object used for creating this instance.
 
-    descriptor_calculator : mala.descriptors.descriptor_base.DescriptorBase
+    descriptor_calculator : mala.descriptors.descriptor.Descriptor
         The descriptor calculator used for parsing/converting fingerprint
         data. If None, the descriptor calculator will be created by this
         object using the parameters provided. Default: None
 
-    target_calculator : mala.targets.target_base.TargetBase
+    target_calculator : mala.targets.target.Target
         Target calculator used for parsing/converting target data. If None,
         the target calculator will be created by this object using the
-        parameters provided.
+        parameters provided. Default: None
 
     Attributes
     ----------
-    descriptor_calculator : mala.descriptors.descriptor_base.DescriptorBase
+    descriptor_calculator : mala.descriptors.descriptor.Descriptor
         Descriptor calculator used for parsing/converting fingerprint data.
 
-    target_calculator : mala.targets.target_base.TargetBase
+    target_calculator : mala.targets.target.Target
         Target calculator used for parsing/converting target data.
     """
 
@@ -44,11 +44,11 @@ class DataConverter:
         self.parameters: ParametersData = parameters.data
         self.target_calculator = target_calculator
         if self.target_calculator is None:
-            self.target_calculator = TargetInterface(parameters)
+            self.target_calculator = Target(parameters)
 
         self.descriptor_calculator = descriptor_calculator
         if self.descriptor_calculator is None:
-            self.descriptor_calculator = DescriptorInterface(parameters)
+            self.descriptor_calculator = Descriptor(parameters)
 
         self.__snapshots_to_convert = []
         self.__snapshot_description = []
@@ -60,7 +60,7 @@ class DataConverter:
         """
         Add a Quantum Espresso snapshot to the list of conversion list.
 
-        Please not that a Quantum Espresso snapshot consists of:
+        Please note that a Quantum Espresso snapshot consists of:
 
             - a Quantum Espresso output file for a self-consistent calculation.
             - multiple .cube files containing the LDOS
@@ -68,10 +68,10 @@ class DataConverter:
         Parameters
         ----------
         qe_out_file : string
-            Name of Quantum Espresso output file for snapshot.
+            Name of Quantum Espresso output file for this snapshot.
 
         qe_out_directory : string
-            Path to Quantum Espresso output file for snapshot.
+            Path to Quantum Espresso output file for this snapshot.
 
         cube_naming_scheme : string
             Naming scheme for the LDOS .cube files.
@@ -108,7 +108,8 @@ class DataConverter:
             Position of the desired snapshot in the snapshot list.
 
         use_memmap : string
-            If not None, a memory mapped file will be used to gather the LDOS.
+            If not None, a memory mapped file with this name will be used to
+            gather the LDOS.
             If run in MPI parallel mode, such a file MUST be provided.
 
         input_path : string
@@ -184,7 +185,8 @@ class DataConverter:
             return tmp_input, tmp_output
 
     def convert_snapshots(self, save_path="./",
-                          naming_scheme="ELEM_snapshot*", starts_at=0):
+                          naming_scheme="ELEM_snapshot*", starts_at=0,
+                          file_based_communication=False):
         """
         Convert the snapshots in the list to numpy arrays.
 
@@ -205,6 +207,12 @@ class DataConverter:
             consistency in naming when converting e.g. only a certain portion
             of all available snapshots. If set to e.g. 4,
             the first snapshot generated will be called snapshot4.
+
+        file_based_communication : bool
+            If True, the LDOS will be gathered using a file based mechanism.
+            This is drastically less performant then using MPI, but may be
+            necessary when memory is scarce. Default is False, i.e., the faster
+            MPI version will be used.
         """
         for i in range(0, len(self.__snapshots_to_convert)):
             snapshot_number = i + starts_at
@@ -213,7 +221,8 @@ class DataConverter:
 
             # A memory mapped file is used as buffer for distributed cases.
             memmap = None
-            if self.parameters._configuration["mpi"]:
+            if self.parameters._configuration["mpi"] and \
+                    file_based_communication:
                 memmap = os.path.join(save_path, snapshot_name +
                                       ".out.npy_temp")
 
@@ -226,5 +235,6 @@ class DataConverter:
                      min_verbosity=0)
 
             if get_rank() == 0:
-                if self.parameters._configuration["mpi"]:
+                if self.parameters._configuration["mpi"] \
+                        and file_based_communication:
                     os.remove(memmap)
